Rename Bullet->BulletData and change its namespace.
